apps/worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lumen-worker"
version = "0.1.0"
description = "Lumen arq Worker — DESIGN.md §6 + §7 上游集成"
requires-python = ">=3.12"
dependencies = [
    "arq>=0.26,<1.0",
    "httpx[socks]>=0.27,<1.0",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "sqlalchemy[asyncio]>=2.0.30,<3.0",
    "asyncpg>=0.29,<1.0",
    "redis>=5.0,<7.0",
    "pillow>=10.3,<13.0",
    "blurhash-python>=1.2,<2.0",
    "lumen-core",
    # 观测层（V1.0 收尾）
    "sentry-sdk>=2.0,<3.0",
    "prometheus-client>=0.20,<1.0",
    "opentelemetry-api>=1.25,<2.0",
    "opentelemetry-sdk>=1.25,<2.0",
    "opentelemetry-exporter-otlp-proto-http>=1.25,<2.0",
    "opentelemetry-instrumentation-sqlalchemy>=0.46b0,<1.0",
    "opentelemetry-instrumentation-httpx>=0.46b0,<1.0",
    "opentelemetry-instrumentation-redis>=0.46b0,<1.0",
]

[tool.uv.sources]
lumen-core = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["app"]
